[Chore] create custom_ftok.h
