+++ custom_ops/gpu_ops/save_output_msg_with_topk.cc
@@ -17,6 +17,7 @@
 #include <sys/ipc.h>
 #include <sys/msg.h>
 #include <sys/types.h>
+#include "custom_ftok.h"
 #include "paddle/extension.h"
 
 #ifndef PD_BUILD_STATIC_OP
